chore(cli/setup): Make setup/deploy ESM compatible
