+++ packages/cli/src/commands/setup/deploy/helpers/helpers.js
@@ -1,0 +1,9 @@
+export function createHandler(componentName) {
+  return async (argv) => {
+    const importedHandler = await import(
+      `../providers/${componentName}Handler.js`
+    )
+
+    return importedHandler(argv)
+  }
+}
